esta todo terrible

+++ source/Entidades/TipoDeDato.cs
@@ -23,7 +23,7 @@
             return nombreUnidadMedida;
         }
 
-        public float getValorUmbral()
+        public double getValorUmbral()
         {
             return valorUmbral;
         }
